Make "Original Aspect Ratio" property available to WebHooks
This property can be populated with a user-defined aspect ratio when the video file's aspect ratio does not match the content's aspect ratio.  This is a common occurrence when video contains encoded black bars (eg UHD blu-ray content).  The "original aspect ratio" is particularly useful in webhooks as it can be used to configure projector lens memory and/or masking systems.